Create class animal, inheritance class animal to create class cow, sheep, goat

+++ BaiTap1 - NET/DataAccessLayer/DataClass/InterfaceAnimal.cs
@@ -18,6 +18,6 @@ namespace BaiTap1___NET.DataAccessLayer.DataClass
 
     public interface IGiveBirth
     {
-        float GiveBirth();
+        int GiveBirth();
     }
 }
